// src/international-space-bar-server/openresponses/generated/zod/responseCompletedStreamingEventSchema.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/

import type { ResponseCompletedStreamingEvent } from "../types/ResponseCompletedStreamingEvent.js";
import { responseResourceSchema } from "./responseResourceSchema.js";
import { z } from "zod/v4";

/**
 * @description A streaming event that indicated the response was completed.
 */
export const responseCompletedStreamingEventSchema = z.object({
    "type": z.enum(["response.completed"]).default("response.completed").describe("The type of the event, always `response.completed`."),
"sequence_number": z.int().describe("The sequence number of the event that was emitted."),
get "response"(){
                return responseResourceSchema.and(z.unknown())
              }
    }).describe("A streaming event that indicated the response was completed.") as unknown as z.ZodType<ResponseCompletedStreamingEvent>

export type ResponseCompletedStreamingEventSchema = ResponseCompletedStreamingEvent