// src/international-space-bar-server/openresponses/generated/zod/responseCreatedStreamingEventSchema.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/

import type { ResponseCreatedStreamingEvent } from "../types/ResponseCreatedStreamingEvent.js";
import { responseResourceSchema } from "./responseResourceSchema.js";
import { z } from "zod/v4";

/**
 * @description A streaming event that indicated the response was created.
 */
export const responseCreatedStreamingEventSchema = z.object({
    "type": z.enum(["response.created"]).default("response.created").describe("The type of the event, always `response.created`."),
"sequence_number": z.int().describe("The sequence number of the event that was emitted."),
get "response"(){
                return responseResourceSchema.and(z.unknown())
              }
    }).describe("A streaming event that indicated the response was created.") as unknown as z.ZodType<ResponseCreatedStreamingEvent>

export type ResponseCreatedStreamingEventSchema = ResponseCreatedStreamingEvent