// src/international-space-bar-server/openresponses/generated/zod/responseOutputItemAddedStreamingEventSchema.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/

import type { ResponseOutputItemAddedStreamingEvent } from "../types/ResponseOutputItemAddedStreamingEvent.js";
import { itemFieldSchema } from "./itemFieldSchema.js";
import { z } from "zod/v4";

/**
 * @description A streaming event that indicated an output item was added to the response.
 */
export const responseOutputItemAddedStreamingEventSchema = z.object({
    "type": z.enum(["response.output_item.added"]).default("response.output_item.added").describe("The type of the event, always `response.output_item.added`."),
"sequence_number": z.int().describe("The sequence number of the event that was emitted."),
"output_index": z.int().describe("The index of the output item that was added."),
get "item"(){
                return z.union([itemFieldSchema.and(z.unknown()), z.null()])
              }
    }).describe("A streaming event that indicated an output item was added to the response.") as unknown as z.ZodType<ResponseOutputItemAddedStreamingEvent>

export type ResponseOutputItemAddedStreamingEventSchema = ResponseOutputItemAddedStreamingEvent