// src/international-space-bar-server/openresponses/generated/zod/responseOutputTextDeltaStreamingEventSchema.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/

import type { ResponseOutputTextDeltaStreamingEvent } from "../types/ResponseOutputTextDeltaStreamingEvent.js";
import { logProbSchema } from "./logProbSchema.js";
import { z } from "zod/v4";

/**
 * @description A streaming event that indicated output text was incrementally added.
 */
export const responseOutputTextDeltaStreamingEventSchema = z.object({
    "type": z.enum(["response.output_text.delta"]).default("response.output_text.delta").describe("The type of the event, always `response.output_text.delta`."),
"sequence_number": z.int().describe("The sequence number of the event that was emitted."),
"item_id": z.string().describe("The ID of the item that was updated."),
"output_index": z.int().describe("The index of the output item that was updated."),
"content_index": z.int().describe("The index of the content part that was updated."),
"delta": z.string().describe("The text delta that was appended."),
get "logprobs"(){
                return z.array(logProbSchema.describe("The log probability of a token.")).describe("The token log probabilities that were emitted with the delta, if any.").optional()
              },
"obfuscation": z.optional(z.string().describe("An obfuscation string that was added to pad the event payload."))
    }).describe("A streaming event that indicated output text was incrementally added.") as unknown as z.ZodType<ResponseOutputTextDeltaStreamingEvent>

export type ResponseOutputTextDeltaStreamingEventSchema = ResponseOutputTextDeltaStreamingEvent